@@ -1,14 +1,12 @@
 import React, { useState, useEffect } from "react";
-import { Card, CardContent, Typography, Snackbar, List, ListItem, IconButton, TextField } from "@material-ui/core";
-import MuiAlert from "@material-ui/lab/Alert";
+import { Card, CardContent, Typography, Snackbar, List, ListItem, IconButton, TextField, Alert } from "@mui/material";
 // import DeleteIcon from "@material-ui/icons/Delete";
-import EmailIcon from "@material-ui/icons/Email";
 import { Chat } from "../../../Types/types";
 import { api } from "../../../Utils/ApiService";
-import { Delete, Visibility, VisibilityOff } from "@material-ui/icons";
+import { Delete, Visibility, VisibilityOff, Email } from "@mui/icons-material";
 import AddDiscordChatClientModal from "../Modals/AddDiscordChatClientModal";
 
-export default function ({ isAdmin }: { isAdmin: boolean }) {
+export default function ChatsCard({ isAdmin }: { isAdmin: boolean }) {
   const [chats, setChats] = useState<Chat[]>([]);
   const [openSuccess, setOpenSuccess] = useState(false);
 
@@ -59,18 +57,12 @@ export default function ({ isAdmin }: { isAdmin: boolean }) {
                 <ListItem key={ch.chatId}>
                   {isAdmin && (
                     <>
-                      <Delete
-                        onClick={() => removeChat(ch.chatId)}
-                        style={{ color: "red", cursor: "pointer", marginRight: "4px" }}
-                      />
+                      <Delete onClick={() => removeChat(ch.chatId)} style={{ color: "red", cursor: "pointer", marginRight: "4px" }} />
                       <IconButton
                         aria-label="toggle visibility"
                         onClick={() =>
                           api.post("/notifications/v1/toggleVisibility", { id: ch.chatId }, null, {
-                            success: () =>
-                              setChats(
-                                chats.map((chat) => (chat.chatId === ch.chatId ? { ...chat, visible: !chat.visible } : chat))
-                              ),
+                            success: () => setChats(chats.map((chat) => (chat.chatId === ch.chatId ? { ...chat, visible: !chat.visible } : chat))),
                             error: () => alert("Chyba"),
                           })
                         }
@@ -79,17 +71,12 @@ export default function ({ isAdmin }: { isAdmin: boolean }) {
                       </IconButton>
                     </>
                   )}
-                  <EmailIcon
-                    onClick={() => testNotification(ch.chatId)}
-                    style={{ cursor: "pointer", marginRight: "24px" }}
-                  ></EmailIcon>
+                  <Email onClick={() => testNotification(ch.chatId)} style={{ cursor: "pointer", marginRight: "24px" }} />
                   {isAdmin ? (
                     <>
                       <TextField
                         value={ch.name}
-                        onChange={(e) =>
-                          setChats(chats.map((chat) => (chat.chatId === ch.chatId ? { ...chat, name: e.target.value } : chat)))
-                        }
+                        onChange={(e) => setChats(chats.map((chat) => (chat.chatId === ch.chatId ? { ...chat, name: e.target.value } : chat)))}
                         onBlur={(e) => api.post("/notifications/v1/rename", { id: ch.chatId, name: ch.name }, null)}
                       />
                     </>
@@ -104,9 +91,9 @@ export default function ({ isAdmin }: { isAdmin: boolean }) {
         </CardContent>
       </Card>
       <Snackbar open={openSuccess} autoHideDuration={6000} onClose={() => setOpenSuccess(false)}>
-        <MuiAlert elevation={6} variant="filled" onClose={() => setOpenSuccess(false)} severity="success">
+        <Alert elevation={6} variant="filled" onClose={() => setOpenSuccess(false)} severity="success">
           Test notification was sent successfully.
-        </MuiAlert>
+        </Alert>
       </Snackbar>
     </>
   );
