@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import App from "./App";
 import "./styles/index.css";
-import { MuiThemeProvider, createMuiTheme } from "@material-ui/core";
+import { createTheme, ThemeProvider } from "@mui/material/styles";
 import * as serviceWorker from "./serviceWorker";
 import { Auth0Provider } from "@auth0/auth0-react";
 
@@ -13,16 +13,28 @@ ReactDOM.render(
     redirectUri={window.location.origin}
     cacheLocation="localstorage"
   >
-    <MuiThemeProvider
-      theme={createMuiTheme({
+    <ThemeProvider
+      theme={createTheme({
         palette: {
           primary: { main: "#343a40" },
           secondary: { main: "#ffffff" },
         },
+        components: {
+          MuiFormControl: {
+            defaultProps: {
+              variant: "standard",
+            },
+          },
+          MuiTextField: {
+            defaultProps: {
+              variant: "standard",
+            },
+          },
+        },
       })}
     >
       <App />
-    </MuiThemeProvider>
+    </ThemeProvider>
   </Auth0Provider>,
   document.getElementById("root")
 );
